Move property parsing to PropertyPlugin
